// src/api/generated/headscale.ts
export interface components {
    schemas: {
// ...
        v1Node: {
            /** Format: uint64 */
            id?: string;
            machineKey?: string;
            nodeKey?: string;
            discoKey?: string;
            ipAddresses?: string[];
            name?: string;
            user?: components["schemas"]["v1User"];
            /** Format: date-time */
            lastSeen?: string;
            /** Format: date-time */
            expiry?: string;
            preAuthKey?: components["schemas"]["v1PreAuthKey"];
            /** Format: date-time */
            createdAt?: string;
            registerMethod?: components["schemas"]["v1RegisterMethod"];
            /**
             * Deprecated
             *     repeated string forced_tags = 18;
             *     repeated string invalid_tags = 19;
             *     repeated string valid_tags = 20;
             */
            givenName?: string;
            online?: boolean;
            approvedRoutes?: string[];
            availableRoutes?: string[];
            subnetRoutes?: string[];
            tags?: string[];
        };
// ...
        v1RegisterMethod: "REGISTER_METHOD_UNSPECIFIED" | "REGISTER_METHOD_AUTH_KEY" | "REGISTER_METHOD_CLI" | "REGISTER_METHOD_OIDC";
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}